+++ UXAssist/Patches/UIPatch.cs
@@ -2,7 +2,6 @@ namespace UXAssist.Patches;
 
 using Common;
 using HarmonyLib;
-using UnityEngine;
 
 [PatchGuid(PluginInfo.PLUGIN_GUID)]
 public class UIPatch: PatchImpl<UIPatch>
@@ -14,10 +13,19 @@ public class UIPatch: PatchImpl<UIPatch>
     }
 
     // Add config button to main menu
-    [HarmonyPostfix, HarmonyPatch(typeof(UIRoot), nameof(UIRoot._OnOpen))]
+    [HarmonyPostfix]
+    [HarmonyPatch(typeof(UIRoot), nameof(UIRoot._OnOpen))]
     public static void UIRoot__OnOpen_Postfix()
     {
         Functions.UIFunctions.InitMenuButtons();
+    }
+
+    [HarmonyPostfix]
+    [HarmonyPatch(typeof(PlanetData), nameof(PlanetData.NotifyScanEnded))]
+    private static void PlanetData_NotifyScanEnded_Postfix(PlanetData __instance)
+    {
+        if (PlanetModelingManager.scnPlanetReqList.Count > 0) return;
+        BepInEx.ThreadingHelper.Instance.StartSyncInvoke(Functions.UIFunctions.OnPlanetScanEnded);
     }
 
     [HarmonyPostfix]